+++ packages/terra-core-docs/src/terra-dev-site/test/card/CardContentCenter.test.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import Card from 'terra-card';
+import CardBody from 'terra-card/lib/CardBody';
+
+export default () => (
+  <Card>
+    <CardBody isContentCentered>
+      Centered content
+    </CardBody>
+  </Card>
+);